move to a proper tesst file
